+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 0.0.1
+# 0.1.0
 
-* TODO: Describe initial release.
+- LinuxLauncher get launcher => _launcher;
+- LinuxDistro get distro => _distro;
+- LinuxSysInfo get sysInfo => _sysInfo;
+- LinuxPower get power => _power;
+- LinuxSudoPrompt get sudoPrompt => _sudoPrompt;
+- LinuxNotify get notify => _noti;
+- LinuxBattery get battery => _battery;
+- LinuxBrightness get brightness => _brightness;
+- LinuxSecurity get security => _security;
+- LinuxSound get sound => _sound;
+- LinuxWifi get wifi => _wifi;